[refactor] split DagsterModel and @record tests
